requirements.txt
pandas
openpyxl
matplotlib
seaborn

langchain
langchain-community
python-dotenv
tabulate

langchain-google-genai
langchain-ollama

langchain-experimental

langgraph

#BD
psycopg2-binary

#checkpoint
langgraph-checkpoint-postgres
psycopg[binary,pool]
